pyproject.toml
[project]
name = "serview"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pyperclip>=1.9.0",
    "pyserial>=3.5",
    "textual>=6.1.0",
]

[project.scripts]
serview = "app:main"

[dependency-groups]
dev = ["ruff>=0.12.12", "textual-dev>=1.7.0"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
include = ["app.py", "widgets"]

[tool.hatch.build.targets.sdist]
include = ["app.py", "widgets"]
